software/user: split library
