+++ litepcie/software/user/Makefile
@@ -1,4 +1,4 @@
-CFLAGS=-O2 -Wall -g -I../kernel -MMD
+CFLAGS=-O2 -Wall -g -I../kernel -Iliblitepcie -MMD
 LDFLAGS=-g
 CC=$(CROSS_COMPILE)gcc
 AR=ar
@@ -7,16 +7,24 @@ PROGS=litepcie_util litepcie_test
 
 all: $(PROGS)
 
-litepcie_util: liblitepcie.o litepcie_util.o
-	$(CC) $(LDFLAGS) -o $@ $^ -lrt -lm
+liblitepcie/liblitepcie.a: liblitepcie/litepcie_dma.o liblitepcie/litepcie_flash.o liblitepcie/litepcie_helpers.o
+	ar rcs $@ $+
+	ranlib $@
 
-litepcie_test: liblitepcie.o litepcie_test.o
-	$(CC) $(LDFLAGS) -o $@ $^ -lrt -lm
+litepcie_util: liblitepcie/liblitepcie.a litepcie_util.o
+	$(CC) $(LDFLAGS) -o $@ $^ -Lliblitepcie -lrt -lm -llitepcie
+
+litepcie_test: liblitepcie/liblitepcie.a litepcie_test.o
+	$(CC) $(LDFLAGS) -o $@ $^ -Lliblitepcie -lrt -lm -llitepcie
 
 clean:
-	rm -f $(PROGS) *.o *.a *.d *~
+	rm -f $(PROGS) *.o *.a *.d *~ liblitepcie/*.a liblitepcie/*.o liblitepcie/*.d
 
 %.o: %.c
 	$(CC) -c $(CFLAGS) -o $@ $<
+
+liblitepcie/%.o: liblitepcie/%.c
+	$(CC) -c $(CFLAGS) -o $@ $<
+
 
 -include $(wildcard *.d)
