+++ litepcie/software/user/liblitepcie/liblitepcie.h
@@ -1,0 +1,18 @@
+/*
+ * LitePCIe library
+ *
+ * This file is part of LitePCIe.
+ *
+ * Copyright (C) 2018-2020 / EnjoyDigital  / florent@enjoy-digital.fr
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#ifndef LITEPCIE_LIB_H
+#define LITEPCIE_LIB_H
+
+#include "litepcie_dma.h"
+#include "litepcie_flash.h"
+#include "litepcie_helpers.h"
+#include "litepcie.h"
+
+#endif /* LITEPCIE_LIB_H */
